feat:修正新增房間價格邏輯(依星期,取消自定義日期對應價格)

+++ src/main/java/com/rafa/sevenTao/service/RoomService.java
@@ -7,7 +7,7 @@ import com.rafa.sevenTao.request.CreateRoomRequest;
 public interface RoomService {
 	public Room createRoom(Hotel hotel, CreateRoomRequest request);
 
-	public boolean deleteRoomByRoomId(int roomId);
+	public boolean deleteRoomByRoomId(long roomId);
 
-	public Room findRoomByRoomId(int roomId);
+	public Room findRoomByRoomId(long roomId);
 }
